MM-69955 Move scroll test helpers into their own file
